@@ -75,7 +75,7 @@ class _SearchBarState extends State<SearchBar> {
                     .where((item) =>
                         item.toLowerCase().contains(query.toLowerCase()))
                     .toList(),
-                overlaySearchListItemBuilder: (item) => Container(
+                overlaySearchListItemBuilder: (dynamic item) => Container(
                       padding: const EdgeInsets.all(8),
                       child: Text(
                         item,
@@ -86,7 +86,7 @@ class _SearchBarState extends State<SearchBar> {
 //                color: Colors.green,
 //                child: Text("no items found..."),
 //              ),
-                onItemSelected: (item) {
+                onItemSelected: (dynamic item) {
                   setState(() {
                     print('$item');
                   });
@@ -133,7 +133,7 @@ class _SearchBarState extends State<SearchBar> {
                     .where((item) =>
                         item.toLowerCase().contains(query.toLowerCase()))
                     .toList(),
-                overlaySearchListItemBuilder: (item) => Container(
+                overlaySearchListItemBuilder: (dynamic item) => Container(
                       padding: const EdgeInsets.all(8),
                       child: Text(
                         item,
@@ -144,7 +144,7 @@ class _SearchBarState extends State<SearchBar> {
 //                color: Colors.green,
 //                child: Text("no items found..."),
 //              ),
-                onItemSelected: (item) {
+                onItemSelected: (dynamic item) {
                   setState(() {
                     print('$item');
                   });
@@ -165,7 +165,7 @@ class _SearchBarState extends State<SearchBar> {
                     .where((item) =>
                         item.toLowerCase().contains(query.toLowerCase()))
                     .toList(),
-                overlaySearchListItemBuilder: (item) => Container(
+                overlaySearchListItemBuilder: (dynamic item) => Container(
                       padding: const EdgeInsets.all(8),
                       child: Text(
                         item,
@@ -176,7 +176,7 @@ class _SearchBarState extends State<SearchBar> {
 //                color: Colors.green,
 //                child: Text("no items found..."),
 //              ),
-                onItemSelected: (item) {
+                onItemSelected: (dynamic item) {
                   setState(() {
                     print('selected item $item');
                   });
